api-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
asyncua==1.0.6
psycopg2-binary==2.9.9
